change average rating to double

+++ LibrarySystem/LibrarySystem/Models/DTO/BookDto.cs
@@ -3,7 +3,7 @@
 namespace LibrarySystem.Models.DTO;
 public class BookDto : BaseBook
 {
-    public int? AverageRating { get; set; }
+    public double? AverageRating { get; set; }
     public int? TotalRatings { get; set; }
     public int? TotalReviews { get; set; }
     public int? TotalViews { get; set; }
